use only one packet processor
